+++ debian/changelog
@@ -1,4 +1,5 @@
   [Michele Simionato]
+  * Extended the taxonomy mapping feature to multiple loss types
   * The error was not stored in the database if the calculation failed
     before starting
   * Made ground_motion_fields=true mandatory in event_based_risk
