+++ openquake/commonlib/tests/logictree_test.py
@@ -2301,7 +2301,7 @@ taxo3,taxo3,1
         with self.assertRaises(openquake.hazardlib.InvalidFile) as ctx:
             inp = dict(taxonomy_mapping=gettemp(xml))
             oq = unittest.mock.Mock(inputs=inp, loss_names=['structural'])
-            logictree.taxonomy_mapping(oq, self.taxonomies)
+            readinput.taxonomy_mapping(oq, self.taxonomies)
         self.assertIn("{'taxo4'} are in the exposure but not in",
                       str(ctx.exception))
 
@@ -2316,7 +2316,7 @@ taxo4,taxo2,.4
         with self.assertRaises(openquake.hazardlib.InvalidFile) as ctx:
             inp = dict(taxonomy_mapping=gettemp(xml))
             oq = unittest.mock.Mock(inputs=inp, loss_names=['structural'])
-            logictree.taxonomy_mapping(oq, self.taxonomies)
+            readinput.taxonomy_mapping(oq, self.taxonomies)
         self.assertIn("the weights do not sum up to 1 for taxo4",
                       str(ctx.exception))
 
@@ -2330,7 +2330,7 @@ taxo4,taxo1,.5
 '''
         inp = dict(taxonomy_mapping=gettemp(xml))
         oq = unittest.mock.Mock(inputs=inp, loss_names=['structural'])
-        lst = logictree.taxonomy_mapping(oq, self.taxonomies)['structural']
+        lst = readinput.taxonomy_mapping(oq, self.taxonomies)['structural']
         self.assertEqual(lst, [[('?', 1)],
                                [('taxo1', 1)],
                                [('taxo2', 1)],
